Add files via upload
This project analyzes the trade relationship between India and Australia, specifically focusing on India’s export and import trends. It calculates the percentage share of India’s exports and imports to/from Australia in 2022 in relation to global trade, as well as the year-on-year growth compared to 2021. The project also employs linear regression techniques to forecast India's exports to Australia for the year 2023. By examining these trade dynamics, the analysis provides valuable insights into the performance and potential growth areas in trade between the two nations, helping stakeholders make informed decisions.